changed userInfo type to NSDictionary*
git-svn-id: svn://orca.physics.unc.edu/Orca@11794 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/ProcessLogic/ProcessElements/analog/rangeTimer/ORRangeTimerModel.h
@@ -45,7 +45,7 @@
 - (void) addAddress;
 - (void) removeAddressAtIndex:(int) anIndex;
 - (void) setEMailList:(NSMutableArray*)aEMailList;
-- (id)   addressEntry:(unsigned)index;
+- (id)   addressEntry:(NSUInteger)index;
 - (BOOL) enableMail;
 - (void) setEnableMail:(BOOL)aEnableMail;
 - (int)	  direction;
@@ -56,7 +56,7 @@
 - (void)  setDeadband:(int)aDeadband;
 - (void) setDeadTimeStart:(NSDate*)aDate;
 - (BOOL) isTrueEndNode;
-- (unsigned) addressCount;
+- (NSUInteger) addressCount;
 
 #pragma mark ¥¥¥Archival
 - (id)   initWithCoder:(NSCoder*)decoder;
